requirements.txt
discord.py-slipstream-fork
aiohttp
python-dotenv
aiosqlite